+++ app/src/main/java/com/example/pictureupload/di/module/FrameworkModule.kt
@@ -1,7 +1,7 @@
 package com.example.pictureupload.di.module
 
 import com.example.pictureupload.data.AuthProvider
-import com.example.pictureupload.framework.AuthProviderImpl
+import com.example.pictureupload.framework.firebase.AuthProviderImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
