+++ app/src/main/java/com/example/pictureupload/framework/firebase/AuthProviderImpl.kt
@@ -1,4 +1,4 @@
-package com.example.pictureupload.framework
+package com.example.pictureupload.framework.firebase
 
 import com.example.pictureupload.data.AuthProvider
 import com.example.pictureupload.domain.AuthResult
